requirement.txt
streamlit==1.30.0
langgraph==0.3.5
langchain-core==0.0.332
langchain-google-genai==0.1.1
python-dotenv==1.0.1
typing-extensions==4.8.0
